refactor: 가용시간 등록 가용시간 수정 api 파일 합치기
